chore: Demo 컴포넌트 분리
[#74]
