+++ packages/frontend/src/components/demo/index.ts
@@ -1,0 +1,3 @@
+import Demo from "./Demo";
+
+export default Demo;
